apps/job_applier_agent/requirements.txt
APScheduler==3.10.4
PyJWT==2.8.0
Flask==3.0.2
Werkzeug==3.0.1
SQLAlchemy==2.0.25
black==24.3.0
pip-tools==7.4.1
pytest==8.0.2
fastapi==0.109.1
httpx==0.27.0
cryptography==42.0.5
pydantic-settings==2.1.0
google-auth==2.27.0
PyPDF2==3.0.1
python-docx==1.1.0